+++ src/bin/command.rs
@@ -15,7 +15,7 @@ use crate::settings::Settings;
 use crate::{WalletFile, WalletPath};
 
 use dusk_wallet::gas::Gas;
-use dusk_wallet::{Address, Dusk, Lux, Wallet};
+use dusk_wallet::{Address, Dusk, Lux, Wallet, EPOCH};
 use dusk_wallet_core::{BalanceInfo, StakeInfo};
 
 /// The default stake gas limit
@@ -360,10 +360,11 @@ impl fmt::Display for RunResult {
             }
             StakeInfo(si, _) => {
                 let stake_str = match si.amount {
-                    Some((value, elegibility)) => format!(
-                        "Current stake amount is: {} DUSK\n> Stake eligibility from block #{}",
+                    Some((value, eligibility)) => format!(
+                        "Current stake amount is: {} DUSK\n> Stake eligibility from block #{} (Epoch {})",
                         Dusk::from(value),
-                        elegibility
+                        eligibility,
+                        eligibility / EPOCH
                     ),
                     None => "No active stake found for this key".to_string(),
                 };